Change to `position: sticky` for header
This change makes the menu scroll with the page until it reaches the top
where it will stick. (In contrast with the current menu which is always
placed a bit from the top.)

On devices where position sticky is not supported, the menu will scroll out
of view, as it does on many sites not using sticky menus. Most browsers do
support it though. Edge users will need version 16 and up.