Update the definition of "A" in JAMstack
Hi JAMstackers! ❤️ 

I'm sending this (very cheeky) PR because I feel that the existing definition of JAMstack is slightly misleading. The _A_ in the _JAM_ acronym is defined as:

> All server-side processes or database actions are abstracted into reusable APIs, accessed over HTTPS with **JavaScript**. These can be custom-built or leverage third-party services.

I see JavaScript as a requirement for all the client-side work, which is indeed a crucial part of the JAMstack architecture. But the _APIs_ part seems to refer to any server-side processing, which can be done with any technology stack and not necessarily JavaScript. I fear that phrasing it like this could induce people in error, making them think that their static site generators, serverless functions or any server-side processing must run on JavaScript/Node.js in order to qualify as a JAMstack project.

I apologise in advance if my interpretation is flawed. If so, please ignore this PR. ➡️🗑 

Thanks!