Add JAMstack Chicago MeetUp
@shortdiv, @prengaiah, and I are launching a JAMstack Chicago MeetUp: https://www.meetup.com/JAMstack-Chicago/

Please add an asset at the artpath. We unanimously agreed on Kapoor’s Cloud Gate sculpture (aka, "The Bean") in black with a reflection of the Chicago skyline in red (i.e., the tower until recently known as the Hancock Tower to the left and the tower forever and always known to Chicagoans as the Sears Tower to the right).