Delete Hyde
Hyde is unmaintained since quite some time and does not work with Python 3. The project has not been archived and do not display a prominent warning of this state, so it seems safer to not propose it to new comers who may spend time on this before running into problems.