Create first headless cms
There are multiple matches for the first headless CMS, including Agility CMS and Contentful: 

Agility CMS
In 2003, Agility CMS developed the first API-driven headless CMS. Agility CMS allows users to integrate their CMS with almost any tool on the market. 

Contentful
Contentful is considered the oldest and best-known headless CMS. Contentful is an API-first headless CMS platform that makes it easy for developers to integrate it with any tech stack.