Remove fresh.deno.dev
According to the homepage:

> Fresh routes are dynamically server-rendered Preact components

I can't find anything in the docs that lets you render to static files
on disk, and the deployment documentation requires Docker or some other
way to run a web server. Fresh looks really neat for building a web
server, but does not support the pre-rendering that's core to the
Jamstack architecture.